# .github/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Atlas Control API
  version: 0.1.0
  description: |
    HTTP API for domain intelligence, graph pivots, and discovery campaigns.
    Workers consume jobs from NATS; CT ingestion runs as a separate service.
servers:
  - url: http://localhost:8090
    description: Local docker-compose stack
paths:
...
  /domains:
    post:
      summary: Seed domains for enrichment
      operationId: seedDomains
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SeedDomainsRequest"
      responses:
        "202":
          description: Enrichment queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SeedDomainsResponse"
  /domains/{domain}:
    get:
      summary: Get domain intelligence
      operationId: getDomain
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: Domain record and edges
        "404":
          description: Domain not found
  /domains/{domain}/subdomains:
    get:
      summary: List subdomains from CT store
      operationId: getDomainSubdomains
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: Subdomain list
  /domains/{domain}/certificates:
    get:
      summary: List certificates for domain
      operationId: getDomainCertificates
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: Certificate list
  /domains/{domain}/rdap:
    get:
      summary: Latest RDAP record
      operationId: getDomainRdap
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: RDAP record or null
  /domains/{domain}/dns:
    get:
      summary: DNS records for domain
      operationId: getDomainDns
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: DNS record list
  /domains/{domain}/pivots:
    get:
      summary: Pivot summary for domain
      operationId: getDomainPivots
      parameters:
        - $ref: "#/components/parameters/domain"
      responses:
        "200":
          description: Available pivot artifacts
  /domains/{domain}/enrich:
    post:
      summary: Re-queue domain enrichment
      operationId: enrichDomain
      parameters:
        - $ref: "#/components/parameters/domain"
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                collectors:
                  type: array
                  items:
                    type: string
      responses:
        "202":
          description: Enrichment queued
...
  /ct/config:
    get:
      summary: Get CT ingestor config
      operationId: getCtConfig
      responses:
        "200":
          description: Ingestor configuration
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CtConfig"
    put:
      summary: Update CT ingestor config
      operationId: setCtConfig
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CtConfig"
      responses:
        "200":
          description: Config updated
  /ct/backfill:
    post:
      summary: Enable TLD-targeted CT backfill
      operationId: startCtBackfill
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CtBackfillRequest"
      responses:
        "202":
          description: Backfill enabled
...
  /campaigns:
    get:
      summary: List campaigns
      operationId: listCampaigns
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
        - name: offset
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Campaign list
    post:
      summary: Create discovery campaign
      operationId: createCampaign
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCampaignRequest"
      responses:
        "201":
          description: Campaign created
...
components:
  parameters:
    domain:
      name: domain
      in: path
      required: true
      schema:
        type: string
        example: example.com
...
  schemas:
    SeedDomainsRequest:
      type: object
      required: [domains]
      properties:
        domains:
          type: array
          items:
            type: string
        collectors:
          type: array
          items:
            type: string
    SeedDomainsResponse:
      type: object
      properties:
        domains:
          type: array
          items:
            type: object
        jobs_queued:
          type: integer
        collectors:
          type: array
          items:
            type: string
...
    CtConfig:
      type: object
      properties:
        target_tlds:
          type: array
          items:
            type: string
        backfill_mode:
          type: boolean
        include_readonly:
          type: boolean
        batches_per_cycle:
          type: integer
        batch_size:
          type: integer
    CtBackfillRequest:
      type: object
      properties:
        target_tlds:
          type: array
          items:
            type: string
        include_readonly:
          type: boolean
        batches_per_cycle:
          type: integer
        batch_size:
          type: integer
    CreateCampaignRequest:
      type: object
      required: [seeds]
      properties:
        seeds:
          type: array
          items:
            type: string
        collectors:
          type: array
          items:
            type: string
        depth:
          type: integer
        limits:
          type: object
          properties:
            max_depth:
              type: integer
            max_entities:
              type: integer
...